bug 修复

+++ src/main/resources/shell/action/initShell.sh
@@ -24,12 +24,16 @@ function init() {
     echo "############copy gz file and xlsx file to computation##############"
     cp "$gz_file_path" "$target_dir/"
     mkdir "$target_dir/test_NGS_data/"
-    tar -zxf "$target_dir/$gz_file_name" -C "$target_dir/test_NGS_data/"
-    # shellcheck disable=SC2045
-    for x in $(ls ./"$target_dir/test_NGS_data/");do
-        mv "$target_dir"/test_NGS_data/"$x"/* "$target_dir"/test_NGS_data
-        rm -rf "$target_dir"/test_NGS_data/"$x"
-    done
+#    tar -zxf "$target_dir/$gz_file_name" -C "$target_dir/test_NGS_data/"
+    unzip -d "$target_dir/test_NGS_data/" "$target_dir/$gz_file_name"
+       # shellcheck disable=SC2045
+     for x in $(ls ./"$target_dir/test_NGS_data/");do
+ #        mv "$target_dir"/test_NGS_data/"$x"/* "$target_dir"/test_NGS_data
+ #        rm -rf "$target_dir"/test_NGS_data/"$x"
+         if [ -d "$target_dir/test_NGS_data/$x" ]; then
+             rm -rf "$target_dir"/test_NGS_data/"$x"
+         fi
+     done
     # shellcheck disable=SC2115
     rm -rf "$target_dir/$gz_file_name"
     #    mv  "$target_dir/test_NGS_data/"
